feat: Deploy dev commands
