+++ src/commands/fun/PhiloInteractionCommand.ts
@@ -1,6 +1,6 @@
 import InteractionCommand from '@structures/command/InteractionCommand';
 import InteractionCommandContext from '@structures/command/InteractionContext';
-import { ApplicationCommandData, MessageAttachment } from 'discord.js-light';
+import { MessageAttachment } from 'discord.js-light';
 import HttpRequests from '@utils/HTTPrequests';
 import { emojis } from '@structures/Constants';
 import { toWritableUTF } from '@utils/Util';
@@ -26,34 +26,6 @@ export default class PhiloInteractionCommand extends InteractionCommand {
   async run(ctx: InteractionCommandContext): Promise<void> {
     const text = ctx.options.getString('frase', true);
     await ctx.defer();
-
-    if (ctx.author.id === '435228312214962204') {
-      const permissionSet: string[] = [];
-
-      const allCommands = ctx.client.slashCommands.reduce<ApplicationCommandData[]>((p, c) => {
-        if (!c.config.devsOnly) return p;
-        permissionSet.push(c.config.name);
-        p.push({
-          name: c.config.name,
-          description: c.config.description,
-          options: c.config.options,
-          defaultPermission: c.config.defaultPermission,
-        });
-        return p;
-      }, []);
-
-      ctx.defer({ content: 'Iniciando deploy' });
-      const res = await ctx.interaction.guild?.commands.set(allCommands);
-
-      res?.forEach((a) => {
-        if (permissionSet.includes(a.name)) {
-          a.permissions.add({
-            permissions: [{ id: ctx.author.id, permission: true, type: 'USER' }],
-          });
-        }
-      });
-      return;
-    }
 
     const res = ctx.client.picassoWs.isAlive
       ? await ctx.client.picassoWs.makeRequest({
